api, metrics: add number of bytes received and sent from/to all entities (#1235)
* API: number of bytes received/sent from/to RTSP connections
* API: number of bytes received/sent from/to RTSP sessions
* API: number of bytes received/sent from/to RTMP connections
* API: number of bytes sent to HLS connections
* API: number of bytes received from paths
* metrics of all the above

+++ internal/rtmp/bytecounter/writer_test.go
@@ -14,5 +14,5 @@ func TestWriter(t *testing.T) {
 	w.SetCount(100)
 
 	w.Write(bytes.Repeat([]byte{0x01}, 64))
-	require.Equal(t, uint32(100+64), w.Count())
+	require.Equal(t, uint64(100+64), w.Count())
 }
